test(duckdb): move tests to specific backend test suites
