+++ ibis/backends/duckdb/tests/snapshots/test_client/test_default_backend/out.sql
@@ -1,0 +1,3 @@
+SELECT
+  SUM(t0.a) AS "Sum(a)"
+FROM ibis_pandas_memtable_fw3sdos5brerlgtmbkopvh334m AS t0
